requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==1.3.5
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-whatsapp-sdk==0.1.0 